refactor(cli): make foundation the bottom layer and enforce it
foundation/ had two upward imports into api/, so the documented layering was
not real and could not be linted.

The template one was a cycle, not just a direction problem:
foundation/config/project.mjs pulled discovery out of api/template, whose
_adapter.mjs imported Project straight back. Project.templates() therefore
triggered a second Project.load() and then discarded the integration templates
that load produced, re-collecting them per-integration for the skip+warn policy.

Neither case was misplaced logic, only misplaced files:

  - api/template/_adapter.mjs had NO api-layer imports — every dependency was
    node:*, jiti, foundation/*, or authoring/*. It is discovery, which
    CONTRIBUTING already assigns to foundation. Moved verbatim to
    foundation/discovery/template-adapter.mjs, which makes the project <-> adapter
    edge intra-layer and leaves api/template importing downward.
  - the on-disk contribution validators (checkRoots + the codemod/template/
    component checks behind validateLoadedIntegration) are shared infra: Project
    collects issues with them and integration-warnings nudges with them. Extracted
    to foundation/integrations/validate-contributions.mjs; api/integration keeps
    the command (validateAtPackageDir and friends) and re-exports
    validateLoadedIntegration so existing importers are unaffected.

With foundation clean, adds the fourth layering rule — foundation/ may not
import api/ or clients/ — split into its own config block because flat config
overrides rather than merges a same-named rule. Verified it fires.

Behavior-preserving: golden 84 invocations, 0 diffs. typecheck:strict 0,
typecheck: authoring 0, 2575 tests pass.
