+++ packages/cli/api/integration/validate-integration.mjs
@@ -30,9 +30,14 @@ import {
   loadManifestObject,
   resolvePackageDir,
 } from '../../foundation/integrations/integrations.mjs';
-import {discoverIntegrationCodemods} from '../../assets/codemods/integration-discovery.mjs';
-import {discoverIntegrationTemplatesForOne} from '../template/template.mjs';
-import * as componentDiscovery from '../../foundation/discovery/component-discovery.mjs';
+// The on-disk contribution validators live in foundation: Project and
+// integration-warnings need them too, and foundation must not depend on api.
+import {
+  validateLoadedIntegration,
+  issueError as error,
+} from '../../foundation/integrations/validate-contributions.mjs';
+
+export {validateLoadedIntegration};
 
 /**
  * @typedef {import('../../foundation/integrations/issue').AstryxIntegrationIssue} Issue
@@ -63,144 +68,6 @@ function findNearestPackageJson(cwd) {
   }
 }
 
-/** @param {string} code @param {string} message @returns {Issue} */
-function error(code, message) {
-  return {code, severity: 'error', message};
-}
-
-/**
- * Verify each declared contribution root exists on disk. A declared-but-missing
- * root is a `missing_root` error.
- * @param {{components?: string, templates?: string, codemods?: string}} resolved
- *   absolute resolved roots (undefined when not declared)
- * @param {Issue[]} issues
- */
-function checkRoots(resolved, issues) {
-  const kinds = /** @type {const} */ (['components', 'templates', 'codemods']);
-  for (const kind of kinds) {
-    const root = resolved[kind];
-    if (root == null) continue;
-    if (!fs.existsSync(root)) {
-      issues.push(
-        error(
-          'missing_root',
-          `Declared ${kind} root does not exist on disk: ${root}`,
-        ),
-      );
-    }
-  }
-}
-
-/**
- * Validate the integration's codemods via the landed discovery. Discovery is
- * strict (throws on bad export / duplicate id); we convert any throw into an
- * `invalid_codemod` error.
- * @param {import('./validate-integration.type.mjs').LoadedIntegration} integration loaded-integration-shaped object
- * @param {Issue[]} issues
- */
-async function checkCodemods(integration, issues) {
-  if (!integration.codemods || !fs.existsSync(integration.codemods)) return;
-  try {
-    await discoverIntegrationCodemods([integration]);
-  } catch (err) {
-    issues.push(error('invalid_codemod', /** @type {any} */ (err).message));
-  }
-}
-
-/**
- * Validate the integration's templates via the landed discovery. Per-template
- * problems are reported as `invalid_template` errors.
- * @param {import('./validate-integration.type.mjs').LoadedIntegration} integration loaded-integration-shaped object
- * @param {Issue[]} issues
- */
-async function checkTemplates(integration, issues) {
-  if (!integration.templates || !fs.existsSync(integration.templates)) return;
-  try {
-    const {errors} = await discoverIntegrationTemplatesForOne(integration);
-    for (const e of errors) {
-      issues.push(error('invalid_template', e.message));
-    }
-  } catch (err) {
-    issues.push(error('invalid_template', /** @type {any} */ (err).message));
-  }
-}
-
-/**
- * Validate the integration's components via the landed ownership discovery.
- * Feature-detected: if the component-ownership export isn't present in this
- * build (sibling PR not yet merged), component validation is skipped rather
- * than hard-failing.
- *
- * `discoverIntegrationComponents` returns ownership records and does not throw
- * on a missing same-stem source — it records `sourcePath: null`. We surface
- * each such record as an `invalid_component` error.
- * @param {import('./validate-integration.type.mjs').LoadedIntegration} integration loaded-integration-shaped object
- * @param {Issue[]} issues
- */
-async function checkComponents(integration, issues) {
-  if (!integration.components || !fs.existsSync(integration.components)) return;
-  const discover = componentDiscovery.discoverIntegrationComponents;
-  if (typeof discover !== 'function') return; // feature not present yet
-  try {
-    const records = (await discover(integration)) ?? [];
-    for (const record of records) {
-      if (record?.sourcePath == null) {
-        issues.push(
-          error(
-            'invalid_component',
-            `Component "${record?.name}" is missing its same-stem source file ${record?.name}.tsx.`,
-          ),
-        );
-      }
-    }
-  } catch (err) {
-    issues.push(error('invalid_component', /** @type {any} */ (err).message));
-  }
-}
-
-/**
- * Run every contribution validator against a loaded-integration-shaped object.
- * @param {import('./validate-integration.type.mjs').LoadedIntegration} integration
- * @param {Issue[]} issues
- */
-async function runContributionChecks(integration, issues) {
-  await checkCodemods(integration, issues);
-  await checkTemplates(integration, issues);
-  await checkComponents(integration, issues);
-}
-
-/**
- * Validate an already-LOADED integration (as produced by
- * `loadIntegrations` in lib/integrations.mjs — absolute contribution roots
- * plus identity) and return its issues. This is the reuse seam for everyday
- * commands that have already loaded the configured integrations and want the
- * SAME validators that `validate-integration` runs, without re-resolving the
- * manifest from disk.
- *
- * The manifest schema is intentionally NOT re-validated here: `loadIntegrations`
- * already validated it (and throws otherwise), so by the time a command holds a
- * loaded integration the manifest is known-good. We re-run the on-disk
- * contribution checks (roots + codemods/templates/components) because those can
- * regress independently of the manifest (a deleted directory, a broken template).
- *
- * @param {import('./validate-integration.type.mjs').LoadedIntegration} loaded loaded-integration-shaped object
- * @returns {Promise<Issue[]>}
- */
-export async function validateLoadedIntegration(loaded) {
-  /** @type {Issue[]} */
-  const issues = [];
-  if (!loaded || typeof loaded !== 'object') return issues;
-  checkRoots(
-    {
-      components: loaded.components,
-      templates: loaded.templates,
-      codemods: loaded.codemods,
-    },
-    issues,
-  );
-  await runContributionChecks(loaded, issues);
-  return issues;
-}
 
 /**
  * Validate a single integration given its package directory and identity.
