+++ packages/cli/api/template/list/list.mjs
@@ -9,11 +9,11 @@
  *   no-skeleton default) here.
  */
 
-import {pkgOf} from '../_adapter.mjs';
+import {pkgOf} from '../../../foundation/discovery/template-adapter.mjs';
 
 /**
  * Project a discovered template set into the `template.list` envelope.
- * @param {import('../_adapter.mjs').DiscoveredTemplate[]} templates
+ * @param {import('../../../foundation/discovery/template-adapter.mjs').DiscoveredTemplate[]} templates
  * @param {{type?: 'page' | 'block', package?: string}} [options]
  * @returns {import('../template.type.mjs').TemplateListResponse}
  */
